Inserted the function checkUserInputInteger und checkUserInputChar into UserInput.h and UserInput.cpp.
